backend/requirements.txt
django==3.2.3
gunicorn==20.1.0
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.0.0
django-shortuuidfield==0.1.3
